+++ modules/effects/shared/src/main/scala/free/effects/implicits.scala
@@ -14,33 +14,7 @@
  * limitations under the License.
  */
 
-package freestyle
+package freestyle.free
 package effects
 
-import cats.mtl.ApplicativeAsk
-
-object reader {
-
-  final class EnvironmentProvider[R] {
-
-    @free abstract class ReaderM {
-      def ask: FS[R]
-      def reader[B](f: R => B): FS[B]
-    }
-
-    trait Implicits {
-
-      implicit def freestyleReaderMHandler[M[_]](
-          implicit AL: ApplicativeAsk[M, R]): ReaderM.Handler[M] =
-        new ReaderM.Handler[M] {
-          def ask: M[R]                  = AL.ask
-          def reader[B](f: R => B): M[B] = AL.reader(f)
-        }
-    }
-
-    object implicits extends Implicits
-  }
-
-  def apply[R] = new EnvironmentProvider[R]
-
-}
+object implicits extends option.Implicits with error.Implicits
