+++ modules/effects/shared/src/main/scala/free/effects/validation.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package freestyle
+package freestyle.free
 package effects
 
 import cats.data.{NonEmptyList, State, Validated, ValidatedNel}
